Move other event types components and add new tests
