+++ src/domain/event/eventInfo/otherEventTimes/otherEventTimes.module.scss
